% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complete_stems}
\alias{complete_stems}
\title{Complete Stems}
\usage{
complete_stems(words, stems)
}
\arguments{
\item{words}{character vector of words}

\item{stems}{character vector of stems}
}
\description{
Complete Stems
}
