% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_words}
\alias{extract_words}
\title{Extract words
Extracts all words from a character string of texts.}
\usage{
extract_words(texts)
}
\arguments{
\item{texts}{character vector of texts}
}
\description{
Extract words
Extracts all words from a character string of texts.
}
\note{
it uses the regex \code{\\b[:alpha:]+\\b} to extract words.
}

