% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{overstemming_index}
\alias{overstemming_index}
\title{Overstemming Index (OI)}
\usage{
overstemming_index(words, stems)
}
\arguments{
\item{words}{is a data.frame containing a column word a a column group
so the function can identify groups of words.}

\item{stems}{is a character vector with the stemming result for each word}
}
\description{
It calculates the proportion of unrelated words that were combined.
}

