% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Performance}
\usage{
performance(stemmers = c("rslp", "hunspell", "porter", "modified-hunspell"))
}
\arguments{
\item{stemmers}{a character vector with names of stemming algorithms.
In the near future, functions will also be accepted.}
}
\value{
a data.frame with the following measures calculated for each stemmer:
\item{UI}{Understemming Index}
\item{OI}{Overstemming Index}
}
\description{
Performance
}
\examples{
\dontrun{perf <- performance()}

}

