% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_porter.R
\name{stem_porter}
\alias{stem_porter}
\title{Stem using Porter's}
\usage{
stem_porter(words, complete = TRUE)
}
\arguments{
\item{words}{character vector of words to be stemmed}

\item{complete}{wheter words must be completed or not (T)

words <- c("balões", "aviões", "avião", "gostou", "gosto", "gostaram")
ptstem:::stem_porter(words)}
}
\description{
This function uses the Porters's algorithm to stem a vector of words.
By default, the Porter's algorithm leaves words cutted. As this makes
reading stemmed texts very difficult, this function provides an option
to complete the stemmed words. By default it completes with the most used
word in the text that has the same stem.
}

