% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{understemming_index}
\alias{understemming_index}
\title{Understemming Index (UI)}
\usage{
understemming_index(words, stems)
}
\arguments{
\item{words}{is a data.frame containing a column word a a column group
so the function can identify groups of words.}

\item{stems}{is a character vector with the stemming result for each word}
}
\description{
It calculates the proportion of related words that had different stems.
}

