% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_hunspell.R
\name{unify_stems}
\alias{unify_stems}
\title{Unify stems by mean position}
\usage{
unify_stems(words, stems)
}
\arguments{
\item{words}{character vector of words}

\item{stems}{character vector of stems}
}
\description{
Hunspell can suggest a list of stems for a word. This function
tries to aggregate all stems into one. Consider the folowing:
}
\details{
a c(1,2)
b c(2,3)
c c(3)

You want that a, b and c to have the same stem.
}
